pyproject.toml
[tool.poetry]
name = "TimeCast"
version = "0.1.0"
description = ""
authors = []



[tool.poetry.dependencies]
python = ">=3.9,<3.10"
numpy = "^1.23.4"
dill = "^0.3.5"
pandas = "^1.5.1"
scipy = "^1.9.3"
scikit-learn = "^1.1.2"
jupyter = "^1.0.0"
ipykernel = "^6.16.0"
matplotlib = "3.5.3"
argparse = "^1.4.0"
importlib = "^1.0.4"
black = "^22.10.0"
tqdm = "^4.64.1"
seaborn = "^0.12.1"
more-itertools = "^9.0.0"
sklearn-pandas = "^2.2.0"
